feat: refactor index page generation and cache response
