+++ external/plugins/swaggerui/impl/ui.go
@@ -3,6 +3,7 @@ package main
 import (
 	"bytes"
 	"embed"
+	"encoding/json"
 	"errors"
 	"fmt"
 	"github.com/imposter-project/imposter-go/external/handler"
@@ -17,6 +18,16 @@ var www embed.FS
 
 var specPrefixPath string
 
+// indexResp is the cached response for the index page.
+var indexResp handler.HandlerResponse
+
+func init() {
+	specPrefixPath = os.Getenv("IMPOSTER_OPENAPI_SPEC_PATH_PREFIX")
+	if specPrefixPath == "" {
+		specPrefixPath = "/_spec"
+	}
+}
+
 // serveStaticContent serves static content from the embedded filesystem.
 func serveStaticContent(path string) handler.HandlerResponse {
 	path = strings.TrimPrefix(path, specPrefixPath)
@@ -24,35 +35,9 @@ func serveStaticContent(path string) handler.HandlerResponse {
 		return handler.HandlerResponse{StatusCode: 404, Body: []byte("File Not Found")}
 	}
 
-	respHeaders := make(map[string]string)
-
 	// index is a special case
 	if path == "/" {
-		// TODO cache this
-		t, err := template.ParseFS(www, "www/index.html.tmpl")
-		if err != nil {
-			return handler.HandlerResponse{
-				StatusCode: 500,
-				Body:       []byte(fmt.Sprintf("error parsing template: %v", err.Error())),
-			}
-		}
-
-		var output bytes.Buffer
-		err = t.Execute(&output, specConfigJSON)
-		if err != nil {
-			return handler.HandlerResponse{
-				StatusCode: 500,
-				Body:       []byte(fmt.Sprintf("error executing template: %v", err.Error())),
-			}
-		}
-		return handler.HandlerResponse{
-			StatusCode: 200,
-			Headers:    respHeaders,
-			Body:       output.Bytes(),
-
-			// used for MIME type detection
-			FileName: "index.html",
-		}
+		return indexResp
 	}
 
 	file, err := www.ReadFile("www" + path)
@@ -68,10 +53,40 @@ func serveStaticContent(path string) handler.HandlerResponse {
 	fileName := path2.Base(path)
 	return handler.HandlerResponse{
 		StatusCode: 200,
-		Headers:    respHeaders,
 		Body:       file,
 
 		// used for MIME type detection
 		FileName: fileName,
 	}
+}
+
+// generateIndexPage generates the index page response using the embedded template.
+func generateIndexPage() error {
+	// serialise the specConfigs to JSON
+	jsonData, err := json.Marshal(specConfigs)
+	if err != nil {
+		return fmt.Errorf("failed to marshal spec config JSON: %w", err)
+	}
+	specConfigJSON := string(jsonData)
+
+	// populate the template with the specConfigJSON
+	t, err := template.ParseFS(www, "www/index.html.tmpl")
+	if err != nil {
+		return fmt.Errorf("error parsing template: %v", err.Error())
+	}
+	var output bytes.Buffer
+	err = t.Execute(&output, specConfigJSON)
+	if err != nil {
+		return fmt.Errorf("error executing template: %v", err.Error())
+	}
+
+	// cache the index response
+	indexResp = handler.HandlerResponse{
+		StatusCode: 200,
+		Body:       output.Bytes(),
+
+		// used for MIME type detection
+		FileName: "index.html",
+	}
+	return nil
 }
